# [perf] Speed up spark creation in sparkline.